Gitchaos/gc.backend/backend/requirements.txt
asgiref==3.2.3
boto3==1.12.5
botocore==1.15.5
coloredlogs==14.0
coverage==5.0.3
Django==3.0.3
django-bulk-sync==1.3.1
django-filter==2.2.0
django-postgres-extra==1.22
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
humanfriendly==7.1.1
jmespath==0.9.4
Markdown==3.2.1
mock==4.0.1
projects==1.0.9
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.25.8
urwid==2.1.0
